backend/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
langchain-groq
groq
google-generativeai
supabase
python-dotenv
pillow
requests
tavily-python
pydantic
python-multipart
